+++ spec/dummy/app/videos/application_video.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class ApplicationVideo < HLS::ApplicationVideo
-  bucket "dummy-bucket"
-  signing_ttl 1800
+  def self.storage = HLS::Storage::S3.new(
+    bucket_name: "dummy-bucket",
+    signing_ttl: 1800
+  )
 end